Workload系の処理を追加。

+++ src/database/migrations/2019_04_28_043645_create_workload_table.php
@@ -17,7 +17,7 @@ class CreateWorkloadTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('project_id');
             $table->unsignedInteger('category_id');
-            $table->float('amount', 2, 2)->comment('the value of workload.');
+            $table->decimal('amount', 4, 2)->comment('the value of workload.');
             $table->date('date');
             $table->timestamps();
         });
